java: Small refactoring

+++ java/benchmark/src/main/java/de/hhu/bsinfo/observatory/benchmark/result/LatencyMeasurement.java
@@ -20,7 +20,7 @@ public class LatencyMeasurement extends Measurement {
         latencyStatistics.stop();
     }
 
-    public void finishMeasuring(double timeInNanos) {
+    public void finishMeasuring(long timeInNanos) {
         totalTime = timeInNanos / 1000000000d;
 
         operationThroughput = (double) getOperationCount() / totalTime;
